ironclad/requirements.txt
flask
torch
torchvision
facenet-pytorch
Pillow
faiss-cpu
numpy